pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-agent"
version = "0.2.0"
requires-python = ">=3.11"
description = "AI-powered Obsidian vault inbox processor"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "watchdog>=4.0",
    "apscheduler>=3.10",
    "openai>=1.30",
    "anthropic>=0.25",
    "google-genai>=1.0",
    "pymupdf>=1.24",
    "markdownify>=0.13",
    "youtube-transcript-api>=0.6",
    "chromadb>=0.5",
    "jinja2>=3.1",
    "anyio>=4.0",
    "python-frontmatter>=1.1",
    "python-dateutil>=2.9",
    "markitdown[docx,pptx,xlsx,xls]>=0.1.0",
]

[project.scripts]
obsidian-agent = "agent.main:cli"

[project.optional-dependencies]
# Audio transcription via Whisper requires PyTorch (large install, enable separately)
audio = [
    "openai-whisper>=20231117",
]
dev = [
    "pytest>=8.0",
    "pytest-anyio>=0.0.0",
    "pytest-httpx>=0.30",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
include = ["agent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require external services (deselect with '-m not integration')",
]
